Add files via upload
Some projects I created in the span of 1 - 2 days. 
The car-test is a car simulator that needs a login and requires you to input a location to work.
The oven_simulator is a smart oven simulator that works like an oven. It also requires a login.
Note:(None of the logins save the information) If you have any further questions or suggestions feel free to contact me.